@@ -11,50 +11,50 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.1.0",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/forms": "~7.1.0",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
-    "apollo-angular": "^1.5.0",
-    "apollo-angular-link-http": "^1.3.1",
-    "apollo-cache-inmemory": "^1.3.2",
-    "apollo-client": "^2.4.0",
-    "apollo-link": "^1.2.0",
-    "apollo-link-context": "^1.0.17",
-    "apollo-link-error": "^1.1.10",
-    "apollo-link-ws": "^1.0.17",
-    "core-js": "^2.5.4",
-    "graphql": "^0.13.2",
+    "@angular/animations": "~8.2.11",
+    "@angular/common": "~8.2.11",
+    "@angular/compiler": "~8.2.11",
+    "@angular/core": "~8.2.11",
+    "@angular/forms": "~8.2.11",
+    "@angular/platform-browser": "~8.2.11",
+    "@angular/platform-browser-dynamic": "~8.2.11",
+    "@angular/router": "~8.2.11",
+    "apollo-angular": "^1.8.0",
+    "apollo-angular-link-http": "^1.9.0",
+    "apollo-cache-inmemory": "^1.6.3",
+    "apollo-client": "^2.6.4",
+    "apollo-link": "^1.2.13",
+    "apollo-link-context": "^1.0.19",
+    "apollo-link-error": "^1.1.12",
+    "apollo-link-ws": "^1.0.19",
+    "core-js": "^3.3.2",
+    "graphql": "^14.5.8",
     "graphql-tag": "^2.10.0",
-    "ngx-weui": "^7.0.0-rc.1",
-    "rxjs": "~6.3.3",
+    "ngx-weui": "^8.0.1",
+    "rxjs": "~6.5.3",
     "subscriptions-transport-ws": "^0.9.16",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.11.0",
-    "@angular/cli": "~7.1.4",
-    "@angular/compiler-cli": "~7.1.0",
-    "@angular/language-service": "~7.1.0",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "@angular-devkit/build-angular": "~0.803.12",
+    "@angular/cli": "~8.3.12",
+    "@angular/compiler-cli": "~8.2.11",
+    "@angular/language-service": "~8.2.11",
+    "@types/jasmine": "~3.4.4",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "~12.11.1",
+    "codelyzer": "~5.1.2",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma": "~4.4.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "ts-node": "~8.4.1",
+    "tslint": "~5.20.0",
+    "typescript": "~3.4.5"
   }
 }
