@@ -131,8 +131,8 @@ export function createApollo(httpLink: HttpLink) {
   const link = split(
     // split based on operation type
     ({ query }) => {
-      const { kind, operation } = getMainDefinition(query);
-      return kind === 'OperationDefinition' && operation === 'subscription';
+      const definition = getMainDefinition(query);
+      return definition.kind === 'OperationDefinition' && definition.operation === 'subscription';
     },
     wsClient,
     httpLink.create({uri}),
